Tests mocker is working I also wrote simple tests for initialize
There is still a problem, that is:
The reference for the lambda functions is the same in the whole execution, so a lambda callback assigned in a test will still be valid in the next one.

This is not exactly a problem, as every test should define their callbacks.

TODO: Maybe define default callbacks. The only return CKR_NOT_IMPLEMENTED or CKR_OK